Add Dataset Splitting notebook
This notebook loads the dataset Liver Tumor Segmentation, thus including all images. After that, the images are divided into different dataframes removing the imbalance between them. A separate train, test and validation for each image type is created from these data frames. After that different folders for train, test and validation are created and the images are filled in them with the annotation.